# docs/api-definition.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Food-Delivery API'
  description: |
    Food-Delivery API based on the OpenAPI 3.0 specification.

    Useful links:
    - [Food-Delivery API Repository](https://github.com/evanigwilo/food-delivery-api)
    - [The API Definition](https://github.com/evanigwilo/food-delivery-api/tree/main/docs/api-definition.yml)
...
paths:
...
  /menu/create:
    post:
      tags:
        - Menu
      summary: Creates a menu
      description: This can only be done by an authenticated user.
      operationId: createMenu
      requestBody:
        required: true
        description: Menu description
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Menu'
      responses:
        '200':
          description: Menu to create was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Menu created successfully.'
                menu:
                  id: '019e24da-dc36-44bc-bbc7-b628731f23a4'
                  name: 'Breakfast Menu'
                  active: true
                  restaurant:
                    id: 'ef138084-a1d3-4a54-b3dc-d587db6400a4'
                    name: 'Tikkaway'
                    rating: 7
        '401':
          $ref: '#/components/responses/Unauthorized'
  /menu/{menuId}:
    get:
      tags:
        - Menu
      summary: Get a menu by id
      security: [] # no authentication
      operationId: getMenu
      parameters:
        - $ref: '#/components/parameters/Menu'
      responses:
        '200':
          description: Returns menu with the specified id
          content:
            application/json:
              schema:
                example:
                  code: 'SUCCESS'
                  message: 'Menu found.'
                  menu:
                    id: '019e24da-dc36-44bc-bbc7-b628731f23a4'
                    name: 'Breakfast Menu'
                    active: true
                    restaurant:
                      id: 'ef138084-a1d3-4a54-b3dc-d587db6400a4'
                      name: 'Tikkaway'
                      rating: 7
        '204':
          description: |
            Menu not found.

            Returns no content.
    patch:
      tags:
        - Menu
      summary: Update menu by id
      description: This can only be done by an authenticated user who created the menu.
      operationId: updateMenu
      requestBody:
        $ref: '#/components/requestBodies/Menu'
      parameters:
        - $ref: '#/components/parameters/Menu'
      responses:
        '200':
          description: Menu to update was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Menu Updated.'

        '400':
          description: Menu to update was not successful
          content:
            application/json:
              examples:
                Failed:
                  $ref: '#/components/examples/MenuNotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Menu
      summary: Delete menu by id
      description: This can only be done by an authenticated user who created the menu.
      operationId: deleteMenu
      parameters:
        - $ref: '#/components/parameters/Menu'
      responses:
        '200':
          description: Menu to delete was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Menu Deleted.'

        '400':
          description: Menu to delete was not successful
          content:
            application/json:
              examples:
                Failed:
                  $ref: '#/components/examples/MenuNotFound'
        '403':
          $ref: '#/components/responses/Unauthorized'

  # Food
...
  /food/create:
    post:
      tags:
        - Food
      summary: Creates a food
      description: This can only be done by an authenticated user.
      operationId: createFood
      requestBody:
        required: true
        description: Food description
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Food'
      responses:
        '200':
          description: Food to create was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Food created successfully.'
                food:
                  id: '241459c5-1f15-4c1a-80b6-36a72314750a'
                  name: 'Small Fresh Tuna'
                  price: 20
                  menu:
                    id: '019e24da-dc36-44bc-bbc7-b628731f23a4'
                    name: 'Breakfast Menu'
                    active: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /food/{foodId}:
    get:
      tags:
        - Food
      summary: Get a food by id
      security: [] # no authentication
      operationId: getFood
      parameters:
        - $ref: '#/components/parameters/Food'
      responses:
        '200':
          description: Returns food with the specified id
          content:
            application/json:
              schema:
                example:
                  code: 'SUCCESS'
                  message: 'Food found.'
                  food:
                    id: '241459c5-1f15-4c1a-80b6-36a72314750a'
                    name: 'Small Fresh Tuna'
                    price: 20
                    menu:
                      id: '019e24da-dc36-44bc-bbc7-b628731f23a4'
                      name: 'Breakfast Menu'
                      active: true
        '204':
          description: |
            Food not found.

            Returns no content.
    patch:
      tags:
        - Food
      summary: Update food by id
      description: This can only be done by an authenticated user who created the food.
      operationId: updateFood
      requestBody:
        $ref: '#/components/requestBodies/Food'
      parameters:
        - $ref: '#/components/parameters/Food'
      responses:
        '200':
          description: Food to update was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Food Updated.'

        '400':
          description: Food to update was not successful
          content:
            application/json:
              examples:
                Failed:
                  $ref: '#/components/examples/FoodNotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Food
      summary: Delete food by id
      description: This can only be done by an authenticated user who created the food.
      operationId: deleteFood
      parameters:
        - $ref: '#/components/parameters/Food'
      responses:
        '200':
          description: Food to delete was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Food Deleted.'

        '400':
          description: Food to delete was not successful
          content:
            application/json:
              examples:
                Failed:
                  $ref: '#/components/examples/FoodNotFound'
        '403':
          $ref: '#/components/responses/Unauthorized'

  # Order
...
components:
  schemas:
...
    Menu:
      type: object
      required:
        - name
        - restaurant
      properties:
        name:
          type: string
          example: 'Room Service Menu'
        restaurant:
          type: string
          format: uuid
          example: 'ef138084-a1d3-4a54-b3dc-d587db6400a4'
    Food:
      type: object
      required:
        - name
        - price
        - menu
      properties:
        name:
          type: string
          example: 'Rice'
        price:
          type: number
          example: 100
        menu:
          type: string
          format: uuid
          example: '019e24da-dc36-44bc-bbc7-b628731f23a4'
...
  examples:
...
    MenuNotFound:
      value:
        code: 'FAILED'
        message: 'Menu not found or User is not creator.'
    FoodNotFound:
      value:
        code: 'FAILED'
        message: 'Food not found or User is not creator.'
...
  responses:
    Unauthorized:
      description: 'User is not authorized'
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: UNAUTHENTICATED
              message:
                type: string
                example: User not authenticated
  parameters:
...
    Menu:
      name: menuId
      in: path
      description: Menu matching the specified id
      required: true
      schema:
        type: string
        format: uuid
    Food:
      name: foodId
      in: path
      description: Food matching the specified id
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    Menu:
      required: false
      description: Menu description
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Menu'
    Food:
      required: false
      description: Food description
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Food'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid # cookie name

# Apply cookie auth globally to all operations
security:
  - cookieAuth: []